requirements.txt
# Core dependencies
meteofrance-api==1.4.0
folium==0.20.0
plotly==6.5.0
pandas==2.3.3
requests==2.32.5
schedule==1.2.2
python-dotenv==1.2.1

# CLI and output
typer==0.20.0
rich==14.2.0

# SMS alerts
twilio==9.3.0
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
gunicorn==23.0.0

# Development/testing
pytest==9.0.2
pytest-cov==7.0.0
